Builtin input functions + lists/arrays - Builtin function 'input_str' - Builtin function 'input_num' - Added Lists/Arrays
# Fixes
- Removed variable declaration on first sight, what was not wished

+++ core/src/Lexer.cpp
@@ -183,7 +183,7 @@ Token Lexer::makeNumber()
 	//std::cout << numStr << std::endl;
 	
 	if (dotCount == 0)
-		return Token(TT_INT, std::stod(numStr), posStart, pos);
+		return Token(TT_INT, std::stoi(numStr), posStart, pos);
 	else
 		return Token(TT_FLOAT, std::stod(numStr), posStart, pos);
 }
